Changed register/create user function
Changed the function to async/await instead of a promise based function.  I was doing it wrong before and now that I learned how to get the ID it is working fine.
Needed to break it down auth().currentUser.uid was not working
Made separate functions   user = awiat auth().currentUser
			  id = await user?.id     		which checks to see if the id is even there first if not it waits for it
								I think that's how it works need to research more on how that works.